Compiler warnings fixed

+++ src/Client/ComponentProxy.cpp
@@ -125,7 +125,7 @@ void ComponentProxy::print(int indent) {
 }
 
 void ComponentProxy::triggerHandler(const std::string & handler) {
-	for (int i = 0; i < handlers.size(); ++i) {
+	for (size_t i = 0; i < handlers.size(); ++i) {
 		if (handlers[i] == handler) {
 			triggerHandler(i);
 			break;
